brew: Further consolidate updated and outdated implementations
