+++ extensions/brew/src/components/actionPanels.tsx
@@ -10,6 +10,7 @@ import {
   type Formula,
   type OutdatedCask,
   type OutdatedFormula,
+  type UpgradePackageStatus,
 } from "../utils";
 import { useTerminalApp } from "../utils/terminal";
 import * as Actions from "./actions";
@@ -354,6 +355,10 @@ export function FormulaActionPanel(props: {
 
 export function OutdatedActionPanel(props: {
   outdated: OutdatedCask | OutdatedFormula;
+  /** Called when the upgrade starts or finishes, e.g. to show its status in a list */
+  onUpgrade?: (status: UpgradePackageStatus) => void;
+  /** Overrides the default "Upgrade All", e.g. to report progress per package */
+  onUpgradeAll?: () => void;
   onAction: (result: boolean) => void;
 }) {
   const { outdated } = props;
@@ -363,12 +368,32 @@ export function OutdatedActionPanel(props: {
     return (o as OutdatedFormula).pinned != undefined;
   }
 
+  // When the caller shows the upgrade status itself, leave the list as-is:
+  // an upgraded package remains visible, with its status
+  function onUpgradeAction(result: boolean) {
+    if (props.onUpgrade) {
+      props.onUpgrade(result ? "upgraded" : "failed");
+    } else {
+      props.onAction(result);
+    }
+  }
+
   return (
     <ActionPanel>
       <ActionPanel.Section>
-        <Actions.FormulaUpgradeAction formula={outdated} onAction={props.onAction} />
-        <Actions.FormulaUpgradeAllAction onAction={props.onAction} />
+        <Actions.FormulaUpgradeAction
+          formula={outdated}
+          onStart={() => props.onUpgrade?.("upgrading")}
+          onAction={onUpgradeAction}
+        />
+        <Actions.FormulaUpgradeAllAction onUpgradeAll={props.onUpgradeAll} onAction={props.onAction} />
         {isPinable(outdated) && <Actions.FormulaPinAction formula={outdated} onAction={props.onAction} />}
+        <Action
+          title="Refresh"
+          icon={Icon.ArrowClockwise}
+          shortcut={Keyboard.Shortcut.Common.Refresh}
+          onAction={() => props.onAction(true)}
+        />
       </ActionPanel.Section>
       <ActionPanel.Section>
         <Action.CopyToClipboard
@@ -398,6 +423,29 @@ export function OutdatedActionPanel(props: {
           onAction={() => runCommandInTerminal(brewUninstallCommand(outdated))}
         />
       </ActionPanel.Section>
+    </ActionPanel>
+  );
+}
+
+/**
+ * Actions available while an upgrade is running.
+ *
+ * Other brew actions are omitted, since Homebrew does not support concurrent processes.
+ */
+export function UpgradingActionPanel(props: { outdated: OutdatedCask | OutdatedFormula; onCancel: () => void }) {
+  return (
+    <ActionPanel>
+      <Action
+        title="Cancel Upgrade"
+        icon={Icon.XMarkCircle}
+        style={Action.Style.Destructive}
+        onAction={props.onCancel}
+      />
+      <Action.CopyToClipboard
+        title="Copy Upgrade Command"
+        content={brewUpgradeCommand(props.outdated)}
+        shortcut={{ modifiers: ["cmd", "opt"], key: "c" }}
+      />
     </ActionPanel>
   );
 }
